+++ src/core/message.interface.ts
@@ -3,4 +3,5 @@ import { EventName } from './event-name.enum';
 export interface Message<T = unknown> {
   name: EventName;
   data?: T;
+  sdkVersion?: string;
 }