+++ src/features/headless-checkout/web-components/apple-pay/apple-pay.component.spec.ts
@@ -28,7 +28,7 @@ describe('ApplePayComponent', () => {
 
     headlessCheckout = {
       form: {
-        onFieldsStatusChange: noopStub,
+        onFieldsStatusChange: () => noopStub,
       },
       events: {
         onCoreEvent: () => noopStub,